Remove unnecessary boxing/unboxing

+++ spring-batch-core/src/main/java/org/springframework/batch/core/launch/support/SimpleJvmExitCodeMapper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2006-2013 the original author or authors.
+ * Copyright 2006-2023 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -31,6 +31,7 @@ import org.springframework.batch.core.ExitStatus;
  * @author Stijn Maller
  * @author Lucas Ward
  * @author Dave Syer
+ * @author Mahmoud Ben Hassine
  */
 
 public class SimpleJvmExitCodeMapper implements ExitCodeMapper {
@@ -81,7 +82,7 @@ public class SimpleJvmExitCodeMapper implements ExitCodeMapper {
 			logger.fatal("Error mapping exit code, generic exit status returned.", ex);
 		}
 
-		return (statusCode != null) ? statusCode.intValue() : JVM_EXITCODE_GENERIC_ERROR;
+		return (statusCode != null) ? statusCode : JVM_EXITCODE_GENERIC_ERROR;
 	}
 
 }
